requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
sse-starlette==1.6.5
aiohttp==3.9.3
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
